+++ src/eval/stmt.rs
@@ -8,12 +8,9 @@ use crate::func::{get_builtin_op_assignment_fn, get_hasher};
 use crate::tokenizer::Token;
 use crate::types::dynamic::{AccessMode, Union};
 use crate::{Dynamic, Engine, RhaiResult, RhaiResultOf, Scope, VarDefInfo, ERR, INT};
+use std::hash::{Hash, Hasher};
 #[cfg(feature = "no_std")]
 use std::prelude::v1::*;
-use std::{
-    convert::TryInto,
-    hash::{Hash, Hasher},
-};
 
 impl Engine {
     /// If the value is a string, intern it.
@@ -307,6 +304,8 @@ impl Engine {
 
                     #[cfg(not(feature = "no_function"))]
                     {
+                        use std::convert::TryInto;
+
                         let rhs_val = self
                             .eval_expr(global, caches, scope, this_ptr.as_deref_mut(), rhs)?
                             .flatten();